refactor(server):previewUserService
